+++ react/ProductPage.js
@@ -1,21 +1,51 @@
 import React, { Component, Fragment } from 'react'
 import PropTypes from 'prop-types'
-import { graphql } from 'react-apollo'
+import { graphql, compose } from 'react-apollo'
 import { ExtensionPoint } from 'render'
 import MicroData from './components/MicroData'
+
+import withDataLayer, { dataLayerProps } from './components/withDataLayer'
 import productQuery from './queries/productQuery.gql'
 
 class ProductPage extends Component {
   static contextTypes = {
     prefetchPage: PropTypes.func,
   }
+
   static propTypes = {
     params: PropTypes.object,
     data: PropTypes.object,
+    ...dataLayerProps,
+  }
+
+  pushToDataLayer = product => {
+    this.props.pushToDataLayer({
+      event: 'productDetail',
+      ecommerce: {
+        detail: {
+          products: [{
+            name: product.productName,
+            brand: product.brand,
+            category: product.categories.length > 0 ? product.categories[0] : undefined,
+            id: product.productId,
+          }],
+        },
+      },
+    })
   }
 
   componentDidMount() {
     this.context.prefetchPage('store/home')
+
+    if (!this.props.data.loading) {
+      this.pushToDataLayer(this.props.data.product)
+    }
+  }
+
+  componentDidUpdate(prevProps) {
+    if (prevProps.data.loading && !this.props.data.loading) {
+      this.pushToDataLayer(this.props.data.product)
+    }
   }
 
   render() {
@@ -23,20 +53,18 @@ class ProductPage extends Component {
     const { loading, variables, product } = data
 
     return (
-      <div>
-        <div className="vtex-product-details-container">
-          {!loading && (
-            <Fragment>
-              <MicroData product={product} />
-              <ExtensionPoint
-                id="container"
-                slug={variables.slug}
-                categories={product.categories}
-                product={product}
-              />
-            </Fragment>
-          )}
-        </div>
+      <div className="vtex-product-details-container">
+        {!loading && (
+          <Fragment>
+            <MicroData product={product} />
+            <ExtensionPoint
+              id="container"
+              slug={variables.slug}
+              categories={product.categories}
+              product={product}
+            />
+          </Fragment>
+        )}
       </div>
     )
   }
@@ -50,6 +78,7 @@ const options = {
   }),
 }
 
-export default graphql(productQuery, options)(
-  ProductPage
-)
+export default compose(
+  graphql(productQuery, options),
+  withDataLayer
+)(ProductPage)